Explicitly declare ierr as an integer variable in the noahmp_init routine
Towards the goal of having no implicitly declared variables in the atmosphere
core, this commit adds a declaration for the 'ierr' variable in the noahmp_init
routine in the mpas_atmphys_lsm_noahmpinit module.

Because the name of the 'ierr' variable begins with the letter 'i', it was
implicitly defined as an integer variable, and its use within the noahmp_init
routine was consistent with that of an integer variable.